moved __init__ contents to client.py
